+++ package.json
@@ -77,7 +77,7 @@
     "jest": "27.3.1",
     "mini-css-extract-plugin": "^2.4.5",
     "nodemon": "^2.0.15",
-    "npm-run-all": "4.1.5",
+    "npm-run-all": "^4.1.5",
     "null-loader": "^4.0.1",
     "postcss": "8.3.11",
     "postcss-advanced-variables": "^3.0.1",
@@ -100,7 +100,7 @@
     "tslib": "^2.3.1",
     "typescript": "4.4.4",
     "url-loader": "^4.1.1",
-    "webpack": "5.61.0",
+    "webpack": "^5.61.0",
     "webpack-cli": "4.9.1",
     "workbox-webpack-plugin": "6.4.1"
   },
